+++ bindings/python/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - [#616]: Fix SentencePiece tokenizers conversion
+- [#616]: Fix offsets produced by Precompiled Normalizer (used by tokenizers converted from SPM)
 
 ## [0.10.0]
 
@@ -298,6 +299,7 @@ delimiter (Works like `.split(delimiter)`)
 - Fix a bug that was causing crashes in Python 3.5
 
 
+[#617]: https://github.com/huggingface/tokenizers/pull/617
 [#616]: https://github.com/huggingface/tokenizers/pull/616
 [#590]: https://github.com/huggingface/tokenizers/pull/590
 [#574]: https://github.com/huggingface/tokenizers/pull/574